fix(style): hackday-conventions-java 의 '4.들여쓰기(indentation)' 규칙에 맞게 수정
- 처음에 'B.7. 커스터마이징' 부분의 내용을 모두 적용하는 바람에 '4.들여쓰기' 규칙을 커스터마이징하여 들여쓰기 기준을 띄어쓰기로 적용해버림.
- '4.들여쓰기(indentation)' 규칙이 적용되도록 'naver-checkstyle-rules.xml' 수정
- [B.7. 커스터마이징](https://naver.github.io/hackday-conventions-java/#checkstyle-customizing)
- [4.들여쓰기(indentation)](https://naver.github.io/hackday-conventions-java/#indentation)
