+++ config/checkstyle/naver-checkstyle-rules.xml
@@ -176,11 +176,11 @@ The following rules in the Naver coding convention cannot be checked by this con
         <property name="tabWidth" value="4"/>
 
         <!-- [indentation-tab] -->
-        <!--<module name="RegexpSinglelineJava">
+        <module name="RegexpSinglelineJava">
             <property name="format" value="^\t* "/>
             <property name="message" value="[indentation-tab] Indent must use tab characters"/>
             <property name="ignoreComments" value="true"/>
-        </module>-->
+        </module>
 
         <!-- End of Indentation chapter -->
 
@@ -418,9 +418,9 @@ The following rules in the Naver coding convention cannot be checked by this con
         </module>
     </module>
 
-    <module name="FileTabCharacter">
+    <!--<module name="FileTabCharacter">
         <property name="eachLine" value="true"/>
-    </module>
+    </module>-->
 
     <!--- Suppression -->
     <!--<module name="SuppressionFilter">
