+++ src/test/java/baseball/ApplicationTest.java
@@ -1,50 +1,50 @@
 package baseball;
 
-import nextstep.test.NSTest;
-import nextstep.utils.Randoms;
+import static org.mockito.Mockito.*;
+
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
 
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.Mockito.mockStatic;
+import nextstep.test.NnSsTest;
+import nextstep.utils.Randoms;
 
-public class ApplicationTest extends NSTest {
-    @BeforeEach
-    void beforeEach() {
-        super.setUp();
-    }
+public class ApplicationTest extends NnSsTest {
+	@BeforeEach
+	void beforeEach() {
+		super.setUp();
+	}
 
-    @Test
-    void 낫싱() {
-        try (final MockedStatic<Randoms> mockRandoms = mockStatic(Randoms.class)) {
-            mockRandoms
-                    .when(() -> Randoms.pickNumberInRange(anyInt(), anyInt()))
-                    .thenReturn(1, 3, 5);
-            running("246");
-            verify("낫싱");
-        }
-    }
+	@Test
+	void nothingTest() {
+		try (final MockedStatic<Randoms> mockRandoms = mockStatic(Randoms.class)) {
+			mockRandoms
+					.when(() -> Randoms.pickNumberInRange(anyInt(), anyInt()))
+					.thenReturn(1, 3, 5);
+			running("246");
+			verify("낫싱");
+		}
+	}
 
-    @Test
-    void 게임종료_후_재시작() {
-        try (final MockedStatic<Randoms> mockRandoms = mockStatic(Randoms.class)) {
-            mockRandoms.when(() -> Randoms.pickNumberInRange(anyInt(), anyInt()))
-                    .thenReturn(7, 1, 3)
-                    .thenReturn(5, 8, 9);
-            run("713", "1", "597", "589", "2");
-            verify("3스트라이크", "게임 끝", "1스트라이크 1볼");
-        }
-    }
+	@Test
+	void stopAndStartGameTest() {
+		try (final MockedStatic<Randoms> mockRandoms = mockStatic(Randoms.class)) {
+			mockRandoms.when(() -> Randoms.pickNumberInRange(anyInt(), anyInt()))
+					.thenReturn(7, 1, 3)
+					.thenReturn(5, 8, 9);
+			run("713", "1", "597", "589", "2");
+			verify("3스트라이크", "게임 끝", "1스트라이크 1볼");
+		}
+	}
 
-    @AfterEach
-    void tearDown() {
-        outputStandard();
-    }
+	@AfterEach
+	void tearDown() {
+		outputStandard();
+	}
 
-    @Override
-    public void runMain() {
-        Application.main(new String[]{});
-    }
+	@Override
+	public void runMain() {
+		Application.main(new String[]{});
+	}
 }
